Invalidate git index cache before smudging
Occasionally we hit the problem where a git smudge filter is not
triggered -- for instance, when a file in a working tree's lstat info
matches the file in the git index cache. (see racy-git.txt in the git
documentation)

The previous work-around was touching the git-fat file; however, in
cases where git fat checkout is automated there is a chance that
touching the file will make no modification to the lstat information
(when a touch happens in the same second as a git checkout). In those
instances, git will not trigger the smudge filter and the file in the
checkout will remain orphaned.

We can work around this by ensuring that the timestamp on a file is
modified by looking at its previous timestamp and adding 1 to it.

Bug: T147856
Change-Id: I9d126b669a71bfd3ce21acee8c3f5d4757e68523